Change the type of `duration_sec` from int to float (#6465)
* style(db): fix typos

* feat(db): make `duration_sec` field in all tables `float64` instead of `int64`

* feat(db): add migration sciprt

* fix(e2e): fix e2e test

+++ backend/core/models/domainlayer/devops/cicd_pipeline.go
@@ -32,7 +32,7 @@ type CICDPipeline struct {
 	Result       string `gorm:"type:varchar(100)"`
 	Status       string `gorm:"type:varchar(100)"`
 	Type         string `gorm:"type:varchar(100);comment: to indicate this is CI or CD"`
-	DurationSec  uint64
+	DurationSec  float64
 	Environment  string `gorm:"type:varchar(255)"`
 	CreatedDate  time.Time
 	FinishedDate *time.Time
